requirements.txt
pytest
pyyaml
pytest-html
pytest-xdist
allure-pytest
pytest-ordering
pytest-rerunfailures
requests